+++ src/test/java/org/morphix/convert/MapConversionsConvertToMapTest.java
@@ -26,7 +26,7 @@ import org.junit.jupiter.api.Test;
  *
  * @author Radu Sebastian LAZIN
  */
-class ToMapTest {
+class MapConversionsConvertToMapTest {
 
 	static class A {